backend/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.5.14
Django==5.1
django-background-tasks==1.2.8
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
facenet-pytorch==2.6.0
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.66.1
h5py==3.11.0
idna==3.7
Jinja2==3.1.4
keras==3.5.0
langchain==0.0.79
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
mtcnn==0.1.1
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.2.0
protobuf==4.25.4
psycopg2==2.9.9
pydantic==1.10.17
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
rich==13.8.0
setuptools==73.0.1
six==1.16.0
SQLAlchemy==1.4.53
sqlparse==0.5.1
sympy==1.13.2
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
